Move certain Syntax querying APIs down from SyntaxGenerator to ISyntaxFacts. These APIs are used by accessibility related analyzers.
This change also requires explicitly linking in the public type DeclarationKind into CodeStyle layer.
