+++ src/Workspaces/Core/Portable/Editing/DeclarationKind.cs
@@ -2,8 +2,11 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-
+#if CODE_STYLE
+namespace Microsoft.CodeAnalysis.Internal.Editing
+#else
 namespace Microsoft.CodeAnalysis.Editing
+#endif
 {
     public enum DeclarationKind
     {
